- Packet updates and small improvements from Tiv - Implemented framework for loading/saving of active filters - Log data is updated after version change and "apply"
git-svn-id: https://svn.code.sf.net/p/dolserver/code/PacketLogConverter/trunk@760 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ LogPackets/CtoS_0xA7_LoginRequest.cs
@@ -73,7 +73,7 @@ namespace PacketLogConverter.LogPackets
 			int minor = ReadByte();
 			int build = ReadByte();
 			int version = major*100 + minor*10 + build;
-			log.Version = version;
+			log.Version = (float)version;
 		}
 
 		/// <summary>
